+++ pypeit/edgetrace.py
@@ -642,9 +642,11 @@ class EdgeTraceSet(DataContainer):
                     table.Column(name='SLITID', dtype=int, length=length,
                                  description='Slit ID Number'),
                     table.Column(name='OBJ_TOPDIST', dtype=float, length=length,
-                                 description='Projected position of the object w.r.t. the top of the slit (arcsec)'),
+                                 description='Projected distance (in arcsec) of the object from the '
+                                             'left edge of the slit (in PypeIt orientation).'),
                     table.Column(name='OBJ_BOTDIST', dtype=float, length=length,
-                                 description='Projected position of the object w.r.t. the bottom of the slit (arcsec)'),
+                                 description='Projected distance (in arcsec) of the object from the '
+                                             'right edge of the slit (in PypeIt orientation)'),
                     table.Column(name='TRACEID', dtype=int, length=length,
                                  description='Row index that matches TRACEID in the design table')
                            ])
@@ -4514,8 +4516,10 @@ class EdgeTraceSet(DataContainer):
         - 'OBJDEC': Declination of the object (deg)
         - 'OBJNAME': Object name assigned by the observer
         - 'SLITID': Slit ID Number (`maskdef_id`)
-        - 'OBJ_TOPDIST': Projected position of the object w.r.t. the top of the slit (arcsec)
-        - 'OBJ_BOTDIST': Projected position of the object w.r.t. the bottom of the slit (arcsec)
+        - 'OBJ_TOPDIST': Projected distance (in arcsec) of the object from the left
+        edge of the slit (in PypeIt orientation)
+        - 'OBJ_BOTDIST': Projected distance (in arcsec) of the object from the right
+        edge of the slit (in PypeIt orientation)
         - 'TRACEID': Row index that matches 'TRACEID' in the design table
 
         Args: