+++ pypeit/spectrographs/__init__.py
@@ -3,6 +3,7 @@ from pypeit.spectrographs import spectrograph
 
 # The import of all the spectrograph modules here is what enables the dynamic
 # compiling of all the available spectrographs below
+from pypeit.spectrographs import gtc_osiris
 from pypeit.spectrographs import bok_bc
 from pypeit.spectrographs import gemini_flamingos
 from pypeit.spectrographs import gemini_gmos
@@ -31,6 +32,7 @@ from pypeit.spectrographs import vlt_fors
 from pypeit.spectrographs import vlt_xshooter
 from pypeit.spectrographs import vlt_sinfoni
 from pypeit.spectrographs import wht_isis
+from pypeit.spectrographs import ntt_efosc2
 
 # Build the list of names for the available spectrographs
 import numpy as np
@@ -54,4 +56,3 @@ def spectrograph_classes():
     return dict([ (c.name,c) for c in spec_c[srt]])
 
 available_spectrographs = list(spectrograph_classes().keys())
-